LayoutTests/http/tests/resourceLoadStatistics/cookies-with-and-without-user-interaction-expected.txt

PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: '<!--frame1-->'
--------
Should receive first-party cookie.
Received cookie named 'firstPartyCookie'.
Client-side document.cookie: firstPartyCookie=value

--------
Frame: '<!--frame2-->'
--------
Should receive no cookies.
Did not receive cookie named 'firstPartyCookie'.
Client-side document.cookie:

--------
Frame: '<!--frame3-->'
--------
After user interaction, should still receive no cookies.
Did not receive cookie named 'firstPartyCookie'.
Client-side document.cookie:
